Eliminate duplicate versions of ``gparams_register_modules.cpp`` ``install_tactic.cpp`` and ``mem_initializer.cpp`` by introducing a component named ``all_init`` that generates these files under the assumption that all other ``LibComponent`` component types are dependencies and should be scanned when generating files.
This implementation is not very satisfactory and is more of a proof of
concept. The problem is that ``all_init.a`` is introduces circular
dependencies (e.g. it depends on ``util.a`` but ``util.a`` also depends
on it). This means that we can't use ``all_init.a`` for linking unless
we specify it on the link line multiple times. To work around this the
object files that constitute ``all_init.a`` are specified on the link
line instead.